Knapsack deals with granularity.

+++ src/main/java/com/algorithms/dp/Item.java
@@ -2,10 +2,10 @@ package com.algorithms.dp;
 
 public class Item {
 	private String name;
-	private int weight;
+	private float weight;
 	private int value;
 	
-	Item(String name, int weight, int value) {
+	Item(String name, float weight, int value) {
 		this.name = name;
 		this.weight = weight;
 		this.value = value;
@@ -15,11 +15,12 @@ public class Item {
 		return name;
 	}
 
-	public int getWeight() {
+	public float getWeight() {
 		return weight;
 	}
 
 	public int getValue() {
 		return value;
 	}
+
 }
